Created a channel manager to help abstract out the twitch client channel functionality. May need more abstractions to fully delineate the channel manager's functionality.
